Composed + MultiplexModelCollection
