+++ Core/Tests/MVVM/MultiplexModelCollectionTests.swift
@@ -8,18 +8,19 @@ class MultiplexModelCollectionTests: XCTestCase {
     func testPassthroughLoadedModels() {
         let first = StaticModel(modelId: "0.0", data: "")
         let second = StaticModel(modelId: "1.0", data: "")
-        let subject = MultiplexModelCollection([
-            StaticModelCollection([[first]]),
-            StaticModelCollection([[second]]),
+
+        let subject = ComposedModelCollection.multiplexing([
+            StaticModelCollection([first]),
+            StaticModelCollection([second]),
             ])
-        let expected: ModelCollectionState = .loaded([[first], [second]])
+        let expected: ModelCollectionState = .loaded([first, second])
         assertModelCollectionState(expected: expected, actual: subject.state)
     }
 
     func testPropegatesLoadingState() {
         let firstSubCollection = SimpleModelCollection()
         let secondSubCollection = SimpleModelCollection()
-        let subject = MultiplexModelCollection([firstSubCollection, secondSubCollection])
+        let subject = ComposedModelCollection.multiplexing([firstSubCollection, secondSubCollection])
         assertModelCollectionState(expected: .notLoaded, actual: subject.state)
         firstSubCollection.onNext(.loading(nil))
         assertModelCollectionState(expected: .loading(nil), actual: subject.state)
@@ -28,7 +29,7 @@ class MultiplexModelCollectionTests: XCTestCase {
     func testPropegatesErrorState() {
         let firstSubCollection = SimpleModelCollection()
         let secondSubCollection = SimpleModelCollection()
-        let subject = MultiplexModelCollection([firstSubCollection, secondSubCollection])
+        let subject = ComposedModelCollection.multiplexing([firstSubCollection, secondSubCollection])
         firstSubCollection.onNext(.loaded([]))
         secondSubCollection.onNext(.error(StubError()))
         assertModelCollectionState(expected: .error(StubError()), actual: subject.state)
@@ -38,14 +39,13 @@ class MultiplexModelCollectionTests: XCTestCase {
         let firstSubCollection = SimpleModelCollection()
         let secondSubCollection = SimpleModelCollection()
         let thirdSubCollection = SimpleModelCollection()
-        let subject = MultiplexModelCollection([firstSubCollection, secondSubCollection, thirdSubCollection])
-        firstSubCollection.onNext(.loaded([[StaticModel(modelId: "0", data: "")]]))
+        let subject = ComposedModelCollection.multiplexing([firstSubCollection, secondSubCollection, thirdSubCollection])
+        firstSubCollection.onNext(.loaded([StaticModel(modelId: "0", data: "")]))
         secondSubCollection.onNext(.loading(nil))
-        thirdSubCollection.onNext(.loaded([[StaticModel(modelId: "1", data: "")]]))
+        thirdSubCollection.onNext(.loaded([StaticModel(modelId: "1", data: "")]))
         let expected: ModelCollectionState = .loading([
-            [StaticModel(modelId: "0", data: "")],
-            [],
-            [StaticModel(modelId: "1", data: "")]
+            StaticModel(modelId: "0", data: ""),
+            StaticModel(modelId: "1", data: "")
             ])
         assertModelCollectionState(expected: expected, actual: subject.state)
     }
@@ -53,12 +53,12 @@ class MultiplexModelCollectionTests: XCTestCase {
     func testPropegatesLoadedState() {
         let firstSubCollection = SimpleModelCollection()
         let secondSubCollection = SimpleModelCollection()
-        let subject = MultiplexModelCollection([firstSubCollection, secondSubCollection])
-        firstSubCollection.onNext(.loaded([[StaticModel(modelId: "0", data: "")]]))
-        secondSubCollection.onNext(.loaded([[StaticModel(modelId: "1", data: "")]]))
+        let subject = ComposedModelCollection.multiplexing([firstSubCollection, secondSubCollection])
+        firstSubCollection.onNext(.loaded([StaticModel(modelId: "0", data: "")]))
+        secondSubCollection.onNext(.loaded([StaticModel(modelId: "1", data: "")]))
         let expected: ModelCollectionState = .loaded([
-            [StaticModel(modelId: "0", data: "")],
-            [StaticModel(modelId: "1", data: "")]
+            StaticModel(modelId: "0", data: ""),
+            StaticModel(modelId: "1", data: "")
             ])
         assertModelCollectionState(expected: expected, actual: subject.state)
     }
